todays final commit

+++ include/core/Room.hpp
@@ -13,19 +13,19 @@ class Room
     
     explicit Room(
         const std::string& roomName, 
-        const std::uint8_t& roomID);
+        const std::uint32_t& roomID);
 
     bool addDevice(std::unique_ptr<Device>&);
     bool removeDevice(std::unique_ptr<Device>&);
     void activateAllDevices();
     void deactivateAllDevices();
-    std::uint8_t getDeviceID(const std::unique_ptr<Device>&) const;
+    std::uint32_t getDeviceID(const std::unique_ptr<Device>&) const;
     std::string getDeviceName(const std::unique_ptr<Device>&) const;
     std::string getDeviceArtNum(const std::unique_ptr<Device>&) const;
     double getDevicePowerUsage(const std::unique_ptr<Device>&) const;
     double getRoomPowerUsage() const;
     std::string getRoomName() const;
-    std::uint8_t getRoomID() const;
+    std::uint32_t getRoomID() const;
 
 
 
@@ -34,7 +34,7 @@ class Room
 
     private:
     std::string roomName_;
-    std::uint8_t roomID_;
+    std::uint32_t roomID_;
     std::vector< std::unique_ptr<Device>> devices_;
 
     ~Room() = default;
